backend/requirements.txt
fastapi
uvicorn
gunicorn
google-cloud-storage
google-cloud-documentai
google-api-core
starlette
python-multipart